[UPDATE] Optimize LazyInMemory provider
