+++ src/Symfony/Bundle/FeatureToggleBundle/DependencyInjection/FeatureToggleExtension.php
@@ -13,12 +13,13 @@ namespace Symfony\Bundle\FeatureToggleBundle\DependencyInjection;
 
 use Symfony\Bundle\FeatureToggleBundle\Strategy\CustomStrategy;
 use Symfony\Component\Config\FileLocator;
+use Symfony\Component\DependencyInjection\Argument\ServiceClosureArgument;
 use Symfony\Component\DependencyInjection\ChildDefinition;
-use Symfony\Component\DependencyInjection\Compiler\ServiceLocatorTagPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Loader\PhpFileLoader;
 use Symfony\Component\DependencyInjection\Reference;
+use Symfony\Component\FeatureToggle\Feature;
 use Symfony\Component\FeatureToggle\Provider\ProviderInterface;
 use Symfony\Component\FeatureToggle\Strategy\StrategyInterface;
 use Symfony\Component\HttpKernel\DependencyInjection\Extension;
@@ -61,20 +62,20 @@ final class FeatureToggleExtension extends Extension
     private function loadFeatures(ContainerBuilder $container, array $config): void
     {
         $features = [];
-        $providerLocatorMap = [];
         foreach ($config['features'] as $featureName => $featureConfig) {
-            $features[$featureName] = [
-                'description' => $featureConfig['description'],
-                'default' => $featureConfig['default'],
-            ];
-            $providerLocatorMap[$featureName] = new Reference($featureConfig['strategy']);
+            $features[$featureName] = new ServiceClosureArgument((new Definition(Feature::class))
+                ->setShared(false)
+                ->setArguments([
+                    $featureName,
+                    $featureConfig['description'],
+                    $featureConfig['default'],
+                    new Reference($featureConfig['strategy']),
+                ]))
+            ;
         }
 
         $container->getDefinition('feature_toggle.provider.lazy_in_memory')
-            ->setArguments([
-                '$features' => $features,
-                '$providerLocator' => ServiceLocatorTagPass::register($container, $providerLocatorMap),
-            ])
+            ->setArgument('$features', $features)
         ;
     }
